Split util.util into util.base and util.common
