+++ message_ix_buildings/chilled/util/base.py
@@ -1,30 +1,8 @@
-import logging
 import os
-from pathlib import Path
 
 import pandas as pd
+from util.common import get_project_root  # type: ignore
 from util.config import Config  # type: ignore
-
-
-def get_project_root() -> Path:
-    return Path(__file__).parent.parent.parent
-
-
-def get_logger(name: str):
-    log = logging.getLogger(name)
-    log.setLevel(logging.INFO)
-
-    # configure the handler and formatter as needed
-    handler = logging.FileHandler(f"{name}.log", mode="w")
-    formatter = logging.Formatter("%(name)s %(asctime)s %(levelname)s %(message)s")
-
-    # add formatter to the handler
-    handler.setFormatter(formatter)
-
-    # add handler to the logger
-    log.addHandler(handler)
-
-    return log
 
 
 def get_archs(config: "Config"):
