+++ message_ix_buildings/chilled/util/common.py
@@ -1,0 +1,23 @@
+import logging
+from pathlib import Path
+
+
+def get_project_root() -> Path:
+    return Path(__file__).parent.parent.parent
+
+
+def get_logger(name: str):
+    log = logging.getLogger(name)
+    log.setLevel(logging.INFO)
+
+    # configure the handler and formatter as needed
+    handler = logging.FileHandler(f"{name}.log", mode="w")
+    formatter = logging.Formatter("%(name)s %(asctime)s %(levelname)s %(message)s")
+
+    # add formatter to the handler
+    handler.setFormatter(formatter)
+
+    # add handler to the logger
+    log.addHandler(handler)
+
+    return log
